requirements.txt
asgiref==3.2.7
Babel==2.8.0
beautifulsoup4==4.9.1
Django==3.0.6
django-bootstrap4==1.1.1
django-phonenumber-field==4.0.0
phonenumbers==8.12.4
psycopg2==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
